+++ packages/presets/src/index.ts
@@ -2,6 +2,7 @@ import { definePreset } from '@pandacss/dev'
 import type { Preset } from '@pandacss/types'
 import { conditions } from './conditions'
 import { globalCss } from './global-css'
+import { breakpoints } from './theme/breakpoints'
 import { keyframes } from './theme/keyframes'
 import { recipes, slotRecipes } from './theme/recipes'
 import { semanticTokens } from './theme/semantic-tokens'
@@ -11,6 +12,7 @@ import { tokens } from './theme/tokens'
 const preset: Preset = definePreset({
   theme: {
     extend: {
+      breakpoints,
       keyframes,
       recipes,
       semanticTokens,