+++ packages/presets/src/theme/tokens/animations.ts
@@ -1,11 +1,5 @@
 import { defineTokens } from '@pandacss/dev'
 
-export const easings = defineTokens.easings({
-  default: { value: 'cubic-bezier(0.2, 0.0, 0, 1.0)' },
-  'emphasized-in': { value: 'cubic-bezier(0.05, 0.7, 0.1, 1.0)' },
-  'emphasized-out': { value: 'cubic-bezier(0.3, 0.0, 0.8, 0.15)' },
-})
-
 export const animations = defineTokens.animations({
   'backdrop-in': {
     value: 'fade-in 250ms {easings.default}',