+++ packages/presets/src/theme/tokens/line-heights.ts
@@ -1,9 +1,0 @@
-import { defineTokens } from '@pandacss/dev'
-
-export const lineHeights = defineTokens.lineHeights({
-  none: { value: '1' },
-  tight: { value: '1.25' },
-  normal: { value: '1.5' },
-  relaxed: { value: '1.75' },
-  loose: { value: '2' },
-})